+++ app/Spago.hs
@@ -16,9 +16,10 @@ import           Spago.Build         (BuildOptions (..), ExtraArg (..), ModuleNa
 import qualified Spago.Build
 import           Spago.GlobalCache   (CacheFlag (..))
 import           Spago.Messages      as Messages
-import           Spago.Packages      (PackageName (..), PackagesFilter (..))
+import           Spago.Packages      (PackageName (..), PackagesFilter (..), JsonFlag(..))
 import qualified Spago.Packages
 import qualified Spago.PscPackage    as PscPackage
+import           Spago.Watch        (ClearScreen (..))
 
 
 -- | Commands that this program handles
@@ -45,7 +46,7 @@ data Command
   | Build BuildOptions
 
   -- | List available packages
-  | ListPackages (Maybe PackagesFilter)
+  | ListPackages (Maybe PackagesFilter) JsonFlag
 
   -- | Verify that a single package is consistent with the Package Set
   | Verify (Maybe Int) (Maybe CacheFlag) PackageName
@@ -110,20 +111,32 @@ parser = do
   command <- projectCommands <|> packageSetCommands <|> pscPackageCommands <|> otherCommands <|> oldCommands
   pure (command, opts)
   where
-    force       = CLI.switch "force" 'f' "Overwrite any project found in the current directory"
-    verbose     = CLI.switch "verbose" 'v' "Enable additional debug logging, e.g. printing `purs` commands"
-    watchBool   = CLI.switch "watch" 'w' "Watch for changes in local files and automatically rebuild"
-    noBuildBool = CLI.switch "no-build" 's' "Skip build step"
+    force           = CLI.switch "force" 'f' "Overwrite any project found in the current directory"
+    verbose         = CLI.switch "verbose" 'v' "Enable additional debug logging, e.g. printing `purs` commands"
+    watchBool       = CLI.switch "watch" 'w' "Watch for changes in local files and automatically rebuild"
+    clearScreenBool = CLI.switch "clear-screen" 'l' "Clear the screen on rebuild (watch mode only)"
+    noBuildBool     = CLI.switch "no-build" 's' "Skip build step"
     watch = do
       res <- watchBool
       pure $ case res of
         True  -> Watch
         False -> BuildOnce
+    clearScreen = do
+      res <- clearScreenBool
+      pure $ case res of
+        True  -> DoClear
+        False -> NoClear
     noBuild = do
       res <- noBuildBool
       pure $ case res of
         True  -> NoBuild
         False -> DoBuild
+    jsonFlagBool = CLI.switch "json" 'j' "Produce JSON output"
+    jsonFlag = do
+      res <- jsonFlagBool
+      pure $ case res of
+        True  -> JsonOutputYes
+        False -> JsonOutputNo
     cacheFlag =
       let wrap = \case
             "skip" -> Just SkipCache
@@ -137,7 +150,7 @@ parser = do
     packageName = CLI.arg (Just . PackageName) "package" "Specify a package name. You can list them with `list-packages`"
     packageNames = CLI.many $ CLI.arg (Just . PackageName) "package" "Package name to add as dependency"
     passthroughArgs = many $ CLI.arg (Just . ExtraArg) " ..any `purs compile` option" "Options passed through to `purs compile`; use -- to separate"
-    buildOptions = BuildOptions <$> limitJobs <*> cacheFlag <*> watch <*> sourcePaths <*> passthroughArgs
+    buildOptions = BuildOptions <$> limitJobs <*> cacheFlag <*> watch <*> clearScreen <*> sourcePaths <*> passthroughArgs
     globalOptions = GlobalOptions <$> verbose
     packagesFilter =
       let wrap = \case
@@ -231,7 +244,7 @@ parser = do
     listPackages =
       ( "list-packages"
       , "List packages available in your packages.dhall"
-      , ListPackages <$> packagesFilter
+      , ListPackages <$> packagesFilter <*> jsonFlag
       )
 
     verify =
@@ -322,7 +335,7 @@ main = do
       Init force                            -> Spago.Packages.initProject force
       Install limitJobs cacheConfig packageNames
         -> Spago.Packages.install limitJobs cacheConfig packageNames
-      ListPackages packagesFilter           -> Spago.Packages.listPackages packagesFilter
+      ListPackages packagesFilter jsonFlag  -> Spago.Packages.listPackages packagesFilter jsonFlag
       Sources                               -> Spago.Packages.sources
       Verify limitJobs cacheConfig package  -> Spago.Packages.verify limitJobs cacheConfig (Just package)
       VerifySet limitJobs cacheConfig       -> Spago.Packages.verify limitJobs cacheConfig Nothing