+++ src/Spago/Build.hs
@@ -41,15 +41,10 @@ data BuildOptions = BuildOptions
   { maybeLimit      :: Maybe Int
   , cacheConfig     :: Maybe GlobalCache.CacheFlag
   , shouldWatch     :: Watch
+  , shouldClear     :: Watch.ClearScreen
   , sourcePaths     :: [Purs.SourcePath]
   , passthroughArgs :: [Purs.ExtraArg]
   }
-
-defaultSourcePaths :: [Purs.SourcePath]
-defaultSourcePaths =
-  [ Purs.SourcePath "src/**/*.purs"
-  , Purs.SourcePath "test/**/*.purs"
-  ]
 
 prepareBundleDefaults
   :: Maybe Purs.ModuleName
@@ -69,7 +64,7 @@ build BuildOptions{..} maybePostBuild = do
   config@Config.Config{ packageSet = PackageSet.PackageSet{..}, ..} <- Config.ensureConfig
   deps <- Packages.getProjectDeps config
   Fetch.fetchPackages maybeLimit cacheConfig deps packagesMinPursVersion
-  let projectGlobs = defaultSourcePaths <> sourcePaths
+  let projectGlobs = configSourcePaths <> sourcePaths
       allGlobs = Packages.getGlobs deps <> projectGlobs
       buildAction = do
         Purs.compile allGlobs passthroughArgs
@@ -79,7 +74,7 @@ build BuildOptions{..} maybePostBuild = do
   absoluteProjectGlobs <- traverse makeAbsolute $ Text.unpack . Purs.unSourcePath <$> projectGlobs
   case shouldWatch of
     BuildOnce -> buildAction
-    Watch     -> Watch.watch (Set.fromAscList $ fmap Glob.compile absoluteProjectGlobs) buildAction
+    Watch     -> Watch.watch (Set.fromAscList $ fmap Glob.compile absoluteProjectGlobs) shouldClear buildAction
 
 -- | Start a repl
 repl :: Spago m => [Purs.SourcePath] -> [Purs.ExtraArg] -> m ()
@@ -87,7 +82,7 @@ repl sourcePaths passthroughArgs = do
   echoDebug "Running `spago repl`"
   config <- Config.ensureConfig
   deps <- Packages.getProjectDeps config
-  let globs = Packages.getGlobs deps <> defaultSourcePaths <> sourcePaths
+  let globs = Packages.getGlobs deps <> Config.configSourcePaths config <> sourcePaths
   Purs.repl globs passthroughArgs
 
 -- | Test the project: compile and run "Test.Main"
@@ -170,4 +165,4 @@ docs sourcePaths = do
   config <- Config.ensureConfig
   deps <- Packages.getProjectDeps config
   echo "Generating documentation for the project. This might take a while.."
-  Purs.docs $ defaultSourcePaths <> Packages.getGlobs deps <> sourcePaths
+  Purs.docs $ Config.configSourcePaths config <> Packages.getGlobs deps <> sourcePaths